% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_missing_processed_data.R
\name{find_missing_processed_data}
\alias{find_missing_processed_data}
\title{Find which soundings are missing from the processed sounding data}
\usage{
find_missing_processed_data(processed_sounding_data)
}
\arguments{
\item{processed_sounding_data}{a list object containing processed sounding data as produced from the 'process_sounding_data' function.}
}
\description{
Find out where expected upper air sounding data are missing.
}

