\name{Process sounding text file into an R list object}
\alias{get.sounding.stations}
\title{
Query the NOAA radiosonde database and create a data frame object that contains a current listing of all sounding stations for which data is available
}

\description{
This function automatically creates a 'df_soundings' R data frame object after querying the NOAA radiosonde database. The function requires no arguments.
}

\usage{
get.sounding.stations()
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

