\name{Process sounding text file into an R list object}
\alias{process.sounding.data}
\title{
Using upper air sounding data, create a list object that can be used for analysis
}

\description{
This function automatically creates a 'sounding_list' R list object from FSL radiosonde data.
}

\usage{
process.sounding.data()
}

\details{

}

\value{

}

\references{

}

\author{
Richard Iannone
}

\note{

}

\seealso{

}

