% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_sounding_data.R
\name{process_sounding_data}
\alias{process_sounding_data}
\title{Process sounding data}
\usage{
process_sounding_data(sounding_data)
}
\arguments{
\item{sounding_data}{a character vector that is produced from the 'get_sounding_data' function.}
}
\description{
Using upper air sounding data, create a list object that can be used for analysis.
}
\examples{
\dontrun{
# Process sounding data created by the 'get.sounding.data' function
processed_data <- process_sounding_data(sounding_data = the_sounding_data)
}
}

