% Generated by roxygen2 (4.0.1): do not edit by hand
\name{process_sounding_stations}
\alias{process_sounding_stations}
\title{Process sounding data}
\usage{
process_sounding_stations(sounding_data)
}
\description{
Using upper air sounding data, create a list object that can be used for analysis.
}
\examples{
\dontrun{
# Process sounding data created by the 'get.sounding.data' function
processed_data <- process_sounding_stations(sounding_data = sounding_data)
}
}

