% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select_sounding_station}
\alias{select_sounding_station}
\title{Search for and select a sounding station}
\usage{
select_sounding_station(stations_df, id_by_wban_wmo = NULL,
  search_station_name = NULL, search_init = NULL, search_wban = NULL,
  search_wmo = NULL, search_prov_state = NULL, search_country = NULL,
  lower_lat = NULL, upper_lat = NULL, lower_lon = NULL,
  upper_lon = NULL, lower_elev = NULL, upper_elev = NULL,
  print_stations = TRUE)
}
\arguments{
\item{stations_df}{a data frame containing identification and location parameters for sounding stations.}

\item{id_by_wban_wmo}{using the combination of WBAN and WMO numbers (with a string in the form of 'XXXXX-YYYYY' in the order of WBAN and WMO), it is possible to uniquely identify a sounding station. When searching with this keyword, all other arguments for which values were supplied will be disregarded.}

\item{search_station_name}{supplying search text in the form of a string will filter the list of available sounding stations and provide feedback on which stations were found. When searching with this keyword, all other arguments for which values were supplied will be disregarded.}

\item{search_init}{}

\item{search_wban}{}

\item{search_wmo}{}

\item{search_prov_state}{}

\item{search_country}{}

\item{lower_lat}{}

\item{upper_lat}{}

\item{lower_lon}{}

\item{upper_lon}{}

\item{lower_elev}{}

\item{upper_elev}{}

\item{print_stations}{if set to TRUE, a returned data frame with 100 or less entries will be printed to the console.}
}
\description{
Filter and select a sounding station from a larger collection.
}
\examples{
\dontrun{
# Get a subset of sounding stations in Canada
sounding_station_subset <- select_sounding_station(search_country = CA)
}
}

