% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.sounding.stations}
\alias{get.sounding.stations}
\title{Get all available sounding stations in dataset}
\usage{
get.sounding.stations()
}
\description{
Query the NOAA radiosonde database and create a data frame object that contains a current listing of all sounding stations for which data is available.
}
\examples{
\dontrun{
# Get the current listing of global soundings as a data frame
all_soundings <- get.sounding.stations()
}
}

