% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_sounding_stations.R
\name{get_sounding_stations}
\alias{get_sounding_stations}
\title{Get all available sounding stations in dataset}
\usage{
get_sounding_stations()
}
\description{
Query the NOAA radiosonde database and create a data frame object that contains a current listing of all sounding stations for which data is available.
}
\examples{
\dontrun{
# Get the current listing of global soundings as a data frame
all_soundings <- get_sounding_stations()
}
}

